[refactor] : 폴더 구조 변경
- 임시 jwt 모듈과 일반 jwt 모듈이 현재 분리되어 있지 않아 사용하기 어려움
- jwt 폴더에 임시 jwt 모듈, 일반 jwt 모듈을 생성
- app.module.ts에 불필요한 import 삭제
- Test code가 작동되게 수정
- jwt 관련 설정 파일 수정
