+++ nestjs/src/auth/auth.service.spec.ts
@@ -1,16 +1,17 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthService } from './auth.service';
 import { PassportsModule } from 'src/passports/passports.module';
-import { TempJwtModule } from 'src/temp-jwt/temp-jwt.module';
+import { TempJwtModule } from 'src/jwt/temp-jwt.module';
 import { MulterModules } from 'src/multer/multer.module';
 import { UserRepository } from './user.repository';
+import { NormalJwtModule } from 'src/jwt/jwt.module';
 
 describe('AuthService', () => {
   let service: AuthService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [PassportsModule, TempJwtModule, MulterModules],
+      imports: [PassportsModule, NormalJwtModule, TempJwtModule, MulterModules],
       providers: [
         AuthService,
         {
